One-pass integration of upstream main (#388 enterprise-concurrency +
env-recipe->CLI levers, #389 file splits/governance, #418, #265) into the
full wip-laguna-lora feature set (Laguna-S/XS, GGUF keep-packed, LoRA v0,
KV/prefix robustness, Nemotron behavior fixes).

Resolution shape (59 conflicted paths):
- Took upstream wholesale where main had absorbed wip work verbatim or
  improved on it: ssm_spill_gate/staging, ssm_snapshot_spill (faultin split),
  bench/agentic/*, residency fault tests (residency_faults.rs), radix spill
  tests (snapshot_reap.rs), dispatch_helpers (GemmDispatch), nvfp4_detect
  (#418), direct_io init-tracking, decode_logits parallel sampling,
  serve.rs/serve_load.rs split, rms_norm.cu symlink.
- Re-applied wip semantics on top of upstream structure: KernelBatchResult +
  up-front radix reservation in the Q12 kernel-batched path, LoRA
  moe_row_adapter/seq_slot metadata + MoeLoraRoute threading (incl. the new
  verify_e/drafter_prefill/speculative sites), grouped-routed-decode
  graph-eager gate + BF16 lm_head batched GEMV in decode_a2 on top of the
  slot-vector graph key, Q2/GGUF keep-packed paths, Laguna mixed
  BF16-shared MoE guards beside upstream's E8M0 guards, thinking/template
  behavior fixes (nano-30b/puzzle MODEL.toml, default_system_prompt,
  cap_thinking_at_max_tokens), stream tool-call salvage beside upstream's
  timeout finish-reason, cwd-hint suppression as a ChatLevers field.
- Wired wip env reads through upstream levers where one exists
  (levers.lora_eager, sched.limits.max_seq_len, SchedLevers watchdogs,
  ctx.dispatch.cublas_gemm); kept env reads only where no lever exists yet
  (ATLAS_PREFILL_VARLEN/CODISPATCH admission, think-loop watchdog,
  ATLAS_NO_DECODE_GRAPHS) - flagged for later per-PR lever conversion.
- Dropped as superseded: wip's ATLAS_SSM_TAIL_REUSE opt-in tail gate (wip
  itself disproved the lookup root cause; upstream's session-gated tail
  lease + insert-overwrite is_tail clearing is the successor), wip's
  ENABLE_LOOP_WATCHDOG statics, wip's rayon sampling commit (on main via
  #388), 92a409b1 nvfp4_detect hunk (superseded by #418).

+++ crates/spark-model/src/layers/moe/forward_prefill_routed.rs
@@ -178,7 +178,7 @@ impl MoeLayer {
         let smallm_nvfp4 = self.experts_scale_kind == crate::weight_map::WeightQuantFormat::Nvfp4;
         let smallm_gate_up = smallm_m > 0
             && total_expanded <= smallm_m
-            && total_expanded as u32 >= fp4_decode_smallm_min()
+            && total_expanded >= fp4_decode_smallm_min()
             && smallm_nvfp4
             && self.moe_fused_gate_up_fp4_smallm.0 != 0
             && self.gate_ptrs_t.is_some()
@@ -471,7 +471,7 @@ impl MoeLayer {
             // null sorted_token_ids (identity row gather), unpermute unchanged.
             let smallm_down = smallm_m > 0
                 && total_expanded <= smallm_m
-                && total_expanded as u32 >= fp4_decode_smallm_min()
+                && total_expanded >= fp4_decode_smallm_min()
                 && smallm_nvfp4
                 && self.moe_down_fp4_smallm.0 != 0
                 && self.down_ptrs_t.is_some();
